refactored fourth p1 function to working status

+++ p1/Models/Tickets.cs
@@ -7,7 +7,7 @@ namespace Models
 {
     public class Tickets
     {
-        public Tickets(float ticketamount, int ticketstatus, string notes, int user_fk, Guid requestid)
+        public Tickets(Decimal ticketamount, int ticketstatus, string notes, int user_fk, Guid requestid)
         {
             TicketAmount = ticketamount;
             TicketStatus = ticketstatus;
@@ -17,7 +17,7 @@ namespace Models
         }
         
         //methods that coincide with the db tickets table's attributes.
-        public float TicketAmount {get; set;}
+        public Decimal TicketAmount {get; set;}
         public int TicketStatus { get; set;}
         public string Notes { get; set;}
         public int User_Fk { get; set;}
